+++ spy/tests/wasm_wrapper.py
